chores: remove loose dependencies and refactor code
