+++ packages/experimental/smart-session/src/client.ts
@@ -1,29 +1,12 @@
 import { ChainController, ConnectionController, OptionsController } from '@reown/appkit-core'
-import { KeyTypes } from './core/helper/index.js'
 import type {
-  Signer,
+  KeyType,
   SmartSessionGrantPermissionsRequest,
   SmartSessionGrantPermissionsResponse,
   WalletGrantPermissionsResponse
 } from './core/utils/TypeUtils.js'
 import { WalletConnectCosigner } from './core/utils/WalletConnectCosigner'
-
-export const ERROR_MESSAGES = {
-  UNSUPPORTED_NAMESPACE: 'Unsupported namespace',
-  NO_RESPONSE_RECEIVED: 'No response received from grantPermissions',
-  INVALID_REQUEST: 'Invalid request structure',
-  INVALID_CHAIN_ID_TYPE: 'Invalid chainId type: must be a string',
-  INVALID_EXPIRY: 'Invalid expiry: must be a positive number',
-  INVALID_PERMISSIONS: 'Invalid permissions: must be a non-empty array',
-  INVALID_POLICIES: 'Invalid policies: must be an array',
-  INVALID_SIGNER: 'Invalid signer: must be an object',
-  INVALID_SIGNER_TYPE: 'Invalid signer type: must be a string',
-  INVALID_KEY_SIGNER: 'A public key is required for key signers',
-  INVALID_KEYS_SIGNER: 'A set of public keys is required for multisig signers',
-  INVALID_ACCOUNT_SIGNER: 'An address is required for account signers',
-  UNSUPPORTED_SIGNER_TYPE: 'Unsupported signer type',
-  INVALID_CHAIN_ID_FORMAT: 'Invalid chainId: must start with "0x"'
-}
+import { ERROR_MESSAGES, validateRequest, validateSigner } from './core/helper/index.js'
 
 // -- Client -------------------------------------------------------------------- //
 // Constants for error messages
@@ -32,7 +15,7 @@ export class AppKitSmartSessionControllerClient {
     request: SmartSessionGrantPermissionsRequest
   ): Promise<SmartSessionGrantPermissionsResponse> {
     try {
-      this.validateRequest(request)
+      validateRequest(request)
       const projectId = OptionsController.state.projectId
       const { activeCaipAddress } = ChainController.state
 
@@ -40,7 +23,7 @@ export class AppKitSmartSessionControllerClient {
         activeCaipAddress && activeCaipAddress.startsWith('eip155:') ? activeCaipAddress : ''
       if (!address) throw new Error(ERROR_MESSAGES.UNSUPPORTED_NAMESPACE)
 
-      this.validateSigner(request.signer)
+      validateSigner(request.signer)
 
       const walletConnectCosigner = new WalletConnectCosigner(projectId)
       const addPermissionResponse = await walletConnectCosigner.addPermission(address, request)
@@ -71,13 +54,13 @@ export class AppKitSmartSessionControllerClient {
     }
   }
 
-  private getCosignerKey(publicKey: `0x${string}`): { type: KeyTypes; publicKey: `0x${string}` } {
-    return { type: KeyTypes.secp256k1, publicKey }
+  private getCosignerKey(publicKey: `0x${string}`): { type: KeyType; publicKey: `0x${string}` } {
+    return { type: 'secp256k1', publicKey }
   }
 
   private updateRequestSigner(
     request: SmartSessionGrantPermissionsRequest,
-    cosignerKey: { type: KeyTypes; publicKey: `0x${string}` }
+    cosignerKey: { type: KeyType; publicKey: `0x${string}` }
   ) {
     const dAppKey = request.signer.type === 'key' ? request.signer.data : undefined
 
@@ -88,64 +71,5 @@ export class AppKitSmartSessionControllerClient {
       }
     }
     return request
-  }
-  private validateRequest(request: SmartSessionGrantPermissionsRequest) {
-    if (typeof request !== 'object' || request === null) {
-      throw new Error(ERROR_MESSAGES.INVALID_REQUEST)
-    }
-    if (typeof request.chainId !== 'string') {
-      throw new Error(ERROR_MESSAGES.INVALID_CHAIN_ID_TYPE)
-    }
-    if (typeof request.chainId !== 'string' || !request.chainId.startsWith('0x')) {
-      throw new Error(ERROR_MESSAGES.INVALID_CHAIN_ID_FORMAT)
-    }
-
-    if (typeof request.expiry !== 'number' || request.expiry <= 0) {
-      throw new Error(ERROR_MESSAGES.INVALID_EXPIRY)
-    }
-
-    if (!Array.isArray(request.permissions) || request.permissions.length === 0) {
-      throw new Error(ERROR_MESSAGES.INVALID_PERMISSIONS)
-    }
-
-    if (!Array.isArray(request.policies)) {
-      throw new Error(ERROR_MESSAGES.INVALID_POLICIES)
-    }
-
-    if (typeof request.signer !== 'object' || request.signer === null) {
-      throw new Error(ERROR_MESSAGES.INVALID_SIGNER)
-    }
-
-    if (typeof request.signer.type !== 'string') {
-      throw new Error(ERROR_MESSAGES.INVALID_SIGNER_TYPE)
-    }
-    if (!['wallet', 'key', 'keys', 'account'].includes(request.signer.type)) {
-      throw new Error(ERROR_MESSAGES.UNSUPPORTED_SIGNER_TYPE)
-    }
-  }
-
-  private validateSigner(signer: Signer) {
-    switch (signer.type) {
-      case 'wallet':
-        // No additional validation needed for wallet signers
-        break
-      case 'key':
-        if (!signer.data.publicKey) {
-          throw new Error(ERROR_MESSAGES.INVALID_KEY_SIGNER)
-        }
-        break
-      case 'keys':
-        if (!signer.data.keys || signer.data.keys.length === 0) {
-          throw new Error(ERROR_MESSAGES.INVALID_KEYS_SIGNER)
-        }
-        break
-      case 'account':
-        if (!signer.data.address) {
-          throw new Error(ERROR_MESSAGES.INVALID_ACCOUNT_SIGNER)
-        }
-        break
-      default:
-        throw new Error(ERROR_MESSAGES.UNSUPPORTED_SIGNER_TYPE)
-    }
   }
 }
